backend/requirements.txt
alabaster==0.7.13
asgiref==3.8.1
babel==2.17.0
backports.zoneinfo==0.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
coverage==7.6.1
django==4.2.27
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
idna==3.11
imagesize==1.4.1
importlib-metadata==8.5.0
jinja2==3.1.6
MarkupSafe==2.1.5
packaging==25.0
psycopg2-binary==2.9.9
pygments==2.19.2
PyJWT==2.9.0
pytz==2025.2
requests==2.32.4
snowballstemmer==3.0.1
sphinx==7.1.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.5.4
typing-extensions==4.13.2
urllib3==2.2.3
zipp==3.20.2
